… PRODUCTION_FEATURES.md
… app/openai_compat.py
… app_enhanced.py
… metrics.py
… tests/conftest.py
… tests/test_enhanced_features.py
… tests/test_performance.py